Replacing new tests for old
