+++ test/mocha.opts
@@ -1,4 +1,4 @@
---require ./test/unit/helpers.js
+--require ./test/helpers.js
 --reporter spec
 --slow 5000
 --exit
