+++ test/opcio.js
@@ -1,105 +1,100 @@
-var chai = require('chai');
-var chaiHttp = require('chai-http');
-const {readFileSync} = require('fs');
-var util = require('util');
-var jwt  = require('jsonwebtoken');
+describe('==== OPCIO ====', () => {
 
-var server = require('../server');
-var config = require('../config');
-var db = require('../modules/db');
-
-var expect = chai.expect;
-var should = chai.should();
-chai.use(chaiHttp);
-
-describe('opcioj', function() {
-    var token = '';
-
-    beforeEach(function(done){
-      var query = util.format('DELETE FROM `opcio`;');
-      db.mysqlExec(query);
-
-      var administranto = {
-        id: 1,
-        uzantnomo: 'nomo',
-        permesoj: [1]
-      };
-      token = jwt.sign(administranto, config.sekretoJWT, {expiresIn: 18000});
+  var token = '';
+  var opcioModel1 = {
+    "priskribo":"priskribo",
+    "idVocxdonado":1
+  }
+  //Before each test we empty the database
+  beforeEach((done) => {
+      createAdmin();
+      cleanTable('opcio');
+      token = generateToken();
       done();
+  });
+
+  describe('GET /opcioj', () => {
+    it('it should POST opcion',(done) => {
+      request
+        .get('/opcioj')
+        .expect(200)
+        .expect((res) => {
+          res.body.length.should.equals(0);
+        })
+      .then((success) => {done()}, (error) => {done(error)});
+    });
+  });
+
+  describe('POST /opcioj', () => {
+    it('it should POST opcion',(done) => {
+      request
+        .post('/opcioj')
+        .set('x-access-token', token)
+        .send(opcioModel1)
+        .expect(201)
+      .then((success) => {done()}, (error) => {done(error)});
     });
 
-    it('it should POST opcion', function(done){
-      chai.request(server)
-          .post('/opcioj')
-          .set('x-access-token', token)
-          .send({"id":1,"priskribo":"priskribo","idVocxdonado":1})
-          .end((err, res) => {
-            res.should.have.status(201);
-            done();
-         });
-     });
+    it('it should NOT POST opcion',(done) => {
+      request
+        .post('/opcioj')
+        .send(opcioModel1)
+        .expect(400)
+      .then((success) => {done()}, (error) => {done(error)});
+    });
+  });
 
-     it('it should NOT POST opcion - sen permeso', function(done){
-       chai.request(server)
-           .post('/opcioj')
-           .send({"id":1,"priskribo":"priskribo","idVocxdonado":1})
-           .end((err, res) => {
-             res.should.have.status(400);
-             done();
-          });
-      });
+  describe('GET /opcioj/:id', () => {
+    it('it should GET opcion - kun opcioj',(done) => {
+      request
+        .post('/opcioj')
+        .set('x-access-token', token)
+        .send(opcioModel1)
+        .expect(201)
+      .then((res) => {
+      return request
+        .get('/opcioj/' + res.body.insertId)
+        .expect(200)
+        .expect((res) => {
+          res.body[0].should.have.property('priskribo');
+          res.body[0].priskribo.should.equal(opcioModel1.priskribo);
+          res.body[0].should.have.property('idVocxdonado');
+          res.body[0].idVocxdonado.should.equal(opcioModel1.idVocxdonado);
+        })
+      })
+      .then((success) => {done()}, (error) => {done(error)});
+    });
+  });
 
-      it('it should GET opcion - sen opcioj', function(done){
-        chai.request(server)
-            .get('/opcioj')
-            .end((err, res) => {
-              res.should.have.status(200);
-              res.body.length.should.equals(0)
-              done();
-           });
-       });
+  describe('DELETE /opcioj/:id', () => {
+    it('it should DELETE opcion',(done) => {
+      request
+        .post('/opcioj')
+        .set('x-access-token', token)
+        .send(opcioModel1)
+        .expect(201)
+      .then((res) => {
+      return request
+        .delete('/opcioj/' + res.body.insertId)
+        .set('x-access-token', token)
+        .expect(204)
+      })
+      .then((success) => {done()}, (error) => {done(error)});
+    });
 
-       it('it should GET opcion - kun opcioj', function(done){
-         var query = util.format('INSERT INTO `opcio` VALUES(1,"priskribo",1);');
-          db.mysqlExec(query).then(function(result){
-             chai.request(server)
-             .get('/opcioj/1')
-             .end((err, res) => {
-               res.should.have.status(200);
-								res.body[0].should.have.property('id');
-								res.body[0].id.should.equal(1);
-								res.body[0].should.have.property('priskribo');
-								res.body[0].priskribo.should.equal("priskribo");
-								res.body[0].should.have.property('idVocxdonado');
-								res.body[0].idVocxdonado.should.equal(1);
-               
-               done();
-             });
-           });
-      });
+    it('it should NOT DELETE opcion - sen permeso',(done) => {
+      request
+        .post('/opcioj')
+        .set('x-access-token', token)
+        .send(opcioModel1)
+        .expect(201)
+      .then((res) => {
+      return request
+        .delete('/opcioj/' + res.body.insertId)
+        .expect(400)
+      })
+      .then((success) => {done()}, (error) => {done(error)});
+    });
+  });
 
-      it('it should DELETE opcion', function(done) {
-        var query = util.format('INSERT INTO `opcio` VALUES(1,"priskribo",1);');
-         db.mysqlExec(query).then(function(result){
-            chai.request(server)
-            .delete('/opcioj/1')
-            .set('x-access-token', token)
-            .end((err, res) => {
-              res.should.have.status(204);
-              done();
-            });
-        });
-      });
-
-      it('it should NOT DELETE opcion - sen permeso', function(done) {
-        var query = util.format('INSERT INTO `opcio` VALUES(1,"priskribo",1);');
-         db.mysqlExec(query).then(function(result){
-            chai.request(server)
-            .delete('/opcioj/1')
-            .end((err, res) => {
-              res.should.have.status(400);
-              done();
-            });
-          });
-      });
 });
